\name{CircOutlier-package}
\alias{CircOutlier-package}
\alias{CircOutlier}
\docType{package}
\title{
 Detecting of Outliers in Circular Regression
}
\description{
Detecting of outliers in circular-circular regression models, modifying its and estimating of models parameters.
}
\details{
\tabular{ll}{
Package: \tab CircOutlier\cr
Type: \tab Package\cr
Version: \tab 3.1.3\cr
Date: \tab 2015-05-16\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Azade Ghazanfarihesari, Majid Sarmad

Maintainer: Azade Ghazanfarihesari <azade_ghazanfari@yahoo.com>
}
\references{
A. H. Abuzaid, A. G. Hussin & I. B. Mohamed (2013) Detecting of outliers in simple circular regression models
using the mean circular error statistics.

Analytical Methods Committe, Robust statistics: a method coping with outliers, Royal Society of Chemistry 2001, amc technical brief. 
}
\seealso{
circular, CircStats
}

