\name{wind}
\alias{wind}
\docType{data}
\title{
Wind Direction
}
\description{
The data used in here, obtained after doing some
calculations on the data to be recorded of Holderness coastline(the Humberside coast of the North sea, UK).
}
\usage{data("wind")}

\references{
A. H. Abuzaid, A. G. Hussin & I. B. Mohamed (2013) Detection of outliers in simple circular   regression models
using the mean circular error statistics.
}


