\name{wind2}
\alias{wind2}
\docType{data}
\title{
Wind Direction
}
\description{
The data used in here, recorded over a period of 22.7 days along the Holderness coastline(the Humberside coast of the North sea, UK) by using 
two different instruments: a high frequency (HF) radar system and an anchored wave buoy.
}
\usage{data("wind2")}

\references{
A. H. Abuzaid, A. G. Hussin & I. B. Mohamed (2013) Detecting of outliers in simple circular   regression models
using the mean circular error statistics.
}
