\name{wind}
\alias{wind}
\docType{data}
\title{
Wind Direction
}
\description{
The data used in here, obtained after doing some
calculations on the data to be recorded of Holderness coastline(the Humberside coast of the North sea, UK).
}
\usage{data("wind")}
\format{
  The format is:
 num [1:122] 0.000337 0.000195 0.000488 0.0005 0.00026 ...
}

\references{
A. H. Abuzaid, A. G. Hussin & I. B. Mohamed (2013) Detecting of outliers in simple circular   regression models
using the mean circular error statistics.
}


